backend/requirements.txt
Django==5.2.6
djangorestframework==3.16.1
django-cors-headers==4.9.0
psycopg2-binary==2.9.10
python-dotenv==1.1.1
djangorestframework-simplejwt==5.5.1
Pillow==11.1.0
